pyproject.toml
[project]
name = "llm-relay"
version = "0.1.0"
description = "Deterministic multi-LLM relay system with strict validation and governance"
authors = [
    {name = "Ethanbrooks23",email = "ethanbrooks@me.com"}
]
requires-python = "^3.12"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "jsonschema (>=4.26.0,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "cryptography (>=46.0.4,<47.0.0)",
    "numpy (>=1.26.0,<3.0.0)",
    "psutil (>=5.9.0,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "anthropic (>=0.80.0,<0.81.0)",
    "apscheduler (>=3.10.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-api-python-client (>=2.0.0,<3.0.0)",
    "google-auth-oauthlib (>=1.0.0,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-analytics-data (>=0.18.0,<1.0.0)",
    "Pillow (>=10.0.0,<12.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "scipy (>=1.12.0,<2.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.30.0,<1.0.0)",
    "websockets (>=12.0,<14.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "openai (>=2.23.0,<3.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "fal-client (>=0.13.1,<0.14.0)",
    "python-telegram-bot[webhooks,rate-limiter] (>=21.0,<22.0.0)",
    "redis (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-asyncio (>=0.24.0,<1.0.0)"
]
cloud-aws = [
    "boto3 (>=1.35.0,<2.0.0)"
]
cloud-gcp = [
    "google-cloud-storage (>=2.14.0,<3.0.0)",
    "google-cloud-run (>=0.10.0,<1.0.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "stochastic: marks tests that verify determinism under repeated/noisy runs (deselect with -m 'not stochastic')",
    "slow: marks tests that are slow (deselect with -m 'not slow')",
]
